fix: missing icons taskbar-mediacontrol
