+++ plugins/utils.ts
@@ -2,7 +2,6 @@ import fs from 'node:fs';
 import path from 'node:path';
 
 import { app, ipcMain, ipcRenderer } from 'electron';
-
 import is from 'electron-is';
 
 import { ValueOf } from '../utils/type-utils';
@@ -14,6 +13,26 @@ export const getMediaIconLocation = () =>
   app.isPackaged
     ? path.resolve(app.getPath('userData'), 'icons')
     : path.resolve(getAssetsDirectoryLocation(), 'media-icons-black');
+
+export const mediaIcons = {
+  play: '\u{1405}', // ᐅ
+  pause: '\u{2016}', // ‖
+  next: '\u{1433}', // ᐳ
+  previous: '\u{1438}', // ᐸ
+};
+
+export const saveMediaIcon = () => {
+  for (const kind of Object.keys(mediaIcons)) {
+    const destinationPath = path.join(getMediaIconLocation(), `${kind}.png`);
+    if (fs.existsSync(destinationPath)) {
+      continue;
+    }
+
+    const iconPath = path.resolve(path.resolve(getAssetsDirectoryLocation(), 'media-icons-black'), `${kind}.png`);
+    fs.mkdirSync(path.dirname(destinationPath), { recursive: true });
+    fs.copyFile(iconPath, destinationPath, () => {});
+  }
+};
 
 // Creates a DOM element from an HTML string
 export const ElementFromHtml = (html: string): HTMLElement => {
